Themis 0.13.0 is now the latest release.

+++ src/wrappers/themis/python/pythemis/scomparator.py
@@ -26,7 +26,7 @@ from .exception import THEMIS_CODES
 
 themis = ctypes.cdll.LoadLibrary(find_library("themis"))
 
-themis.secure_comparator_get_result.restype = ctypes.c_int64
+themis.secure_comparator_get_result.restype = ctypes.c_int32
 
 scomparator_create = themis.secure_comparator_create
 scomparator_create.restype = ctypes.POINTER(ctypes.c_int)
